requirements.txt
fastapi
uvicorn
invisible-watermark
Pillow
python-multipart